# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /users:
    get:
      tags:
        - user-controller
      description: List of all users in DB
      operationId: getAll
      responses:
        "403":
          description: "For user: User cannot access this function."
        "200":
          description: "For admin: List of all users in DB (JSON array)."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockUser'
      security:
        - Bearer Authentication: []
    put:
      tags:
        - user-controller
      description: Update a user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockUser'
        required: true
      responses:
        "200":
          description: User is successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockUser'
        "403":
          description: "For user: If user wants to PUT for another user."
        "404":
          description: "For administrator: No such user."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      security:
        - Bearer Authentication: []
    post:
      tags:
        - user-controller
      description: Creating a user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockUser'
        required: true
      responses:
        "201":
          description: "For admin: User is successfully created in DB "
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockUser'
        "403":
          description: "For user: If user wants to create."
        "400":
          description: "For administrator: User should have first name."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      security:
        - Bearer Authentication: []
...
  /users/{userId}:
    get:
      tags:
        - user-controller
      description: Get user with particular ID.
      operationId: getById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: If user is found in DB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockUser'
        "404":
          description: "For administrator: No such stockUser in DB."
          content:
            application.json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "403":
          description: "For user: If user wants to access another user."
      security:
        - Bearer Authentication: []
    delete:
      tags:
        - user-controller
      description: Deleting a User
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "403":
          description: "For user: If user wants to delete another user."
        "404":
          description: "For administrator: No such user."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "204":
          description: User is successfully deleted
      security:
        - Bearer Authentication: []
  /users/me:
    get:
      tags:
        - user-controller
      operationId: getSelf
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
      security:
        - Bearer Authentication: []
...
components:
  schemas:
    StockUser:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        firstName:
          type: string
        secondName:
          type: string
        birthday:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: boolean
    ErrorModel:
      type: object
      properties:
        error:
          type: string
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT